Barcode clock-in
Allow barcode clock-in. This required a few changes:

  - Add a punchnext field to punchlist table which specifies which punch
    status should be moved to when we scan our barcode.

  - Add a barcode UNIQUE field to the employees table.

  - Modify user search to accept barcodes: additionally, I removed the
    restriction of searching on only a single search term. Terms are now
    AND-ed together.

  - Create new configuration options "$barcode_clockin" and
    "$manual_clockin" which toggle the display of each login form on the
    main clock-in screen.

Due to the creation of the "punchnext" property, the In/Out punch field no
longer needs to be filled in even for normal punches (provided the
"punchnext" fields have been populated by an administrator).

The barcode entry is configured to not auto-complete and is automatically
focused on page load, so any barcode scanner or card reader which emulates
keyboard input should work fine.